move formatFailureMessage to utils
