+++ src/lib/utils.js
@@ -7,6 +7,7 @@
  */
 'use strict';
 
+var colors = require('./colors');
 var fs = require('graceful-fs');
 var path = require('path');
 var Q = require('q');
@@ -319,6 +320,59 @@ function runContentWithLocalBindings(contextRunner, scriptContent, scriptPath,
   }
 }
 
+/**
+ * Given a test result, return a human readable string representing the
+ * failures.
+ *
+ * @param {Object} testResult
+ * @param {boolean} color true if message should include color flags
+ * @return {String}
+ */
+function formatFailureMessage(testResult, color) {
+  var colorize = color ? colors.colorize : function (str) { return str; };
+  var ancestrySeparator = ' \u203A ';
+  var descBullet = colorize('\u25cf ', colors.BOLD);
+  var msgBullet = '  - ';
+  var msgIndent = msgBullet.replace(/./g, ' ');
+
+  return testResult.testResults.filter(function (result) {
+    return result.failureMessages.length !== 0;
+  }).map(function(result) {
+    var failureMessages = result.failureMessages.map(function (errorMsg) {
+      // Filter out q and jasmine entries from the stack trace.
+      // They're super noisy and unhelpful
+      errorMsg = errorMsg.split('\n').filter(function(line) {
+        if (/^\s+at .*?/.test(line)) {
+          // Extract the file path from the trace line
+          var filePath = line.match(/(?:\(|at (?=\/))(.*):[0-9]+:[0-9]+\)?$/);
+          if (filePath
+              && STACK_TRACE_LINE_IGNORE_RE.test(filePath[1])) {
+            return false;
+          }
+        }
+        return true;
+      }).join('\n');
+
+      return msgBullet + errorMsg.replace(/\n/g, '\n' + msgIndent);
+    }).join('\n');
+
+    var testTitleAncestry = result.ancestorTitles.map(function(title) {
+      return colorize(title, colors.BOLD);
+    }).join(ancestrySeparator) + ancestrySeparator;
+
+    return descBullet + testTitleAncestry + result.title + '\n' +
+      failureMessages;
+  }).join('\n');
+}
+
+// A RegExp that matches paths that should not be included in error stack traces
+// (mostly because these paths represent noisy/unhelpful libs)
+var STACK_TRACE_LINE_IGNORE_RE = new RegExp('^(?:' + [
+    path.resolve(__dirname, '..', 'node_modules', 'q'),
+    path.resolve(__dirname, '..', 'vendor', 'jasmine')
+].join('|') + ')');
+
+
 exports.escapeStrForRegex = escapeStrForRegex;
 exports.getLineCoverageFromCoverageInfo = getLineCoverageFromCoverageInfo;
 exports.getLinePercentCoverageFromCoverageInfo =
@@ -329,3 +383,4 @@ exports.normalizeConfig = normalizeConfig;
 exports.pathNormalize = pathNormalize;
 exports.readAndPreprocessFileContent = readAndPreprocessFileContent;
 exports.runContentWithLocalBindings = runContentWithLocalBindings;
+exports.formatFailureMessage = formatFailureMessage;
